+++ Foxtaur/Foxtaur.LibRenderer/Services/Abstractions/Camera/ICamera.cs
@@ -5,9 +5,9 @@ namespace Foxtaur.LibRenderer.Services.Abstractions.Camera;
 
 public class OnZoomChangedArgs
 {
-    public float Zoom { get; private set; }
+    public double Zoom { get; private set; }
 
-    public OnZoomChangedArgs(float zoom)
+    public OnZoomChangedArgs(double zoom)
     {
         Zoom = zoom;
     }
@@ -21,17 +21,17 @@ public interface ICamera
     /// <summary>
     /// Camera latitude
     /// </summary>
-    float Lat { get; set; }
+    double Lat { get; set; }
 
     /// <summary>
     /// Camera longitude
     /// </summary>
-    float Lon { get; set; }
+    double Lon { get; set; }
 
     /// <summary>
     /// Camera height (over the center of Earth)
     /// </summary>
-    float H { get; set; }
+    double H { get; set; }
 
     /// <summary>
     /// Camera position for given geo coordinates
@@ -46,7 +46,7 @@ public interface ICamera
     /// <summary>
     /// Camera zoom (in radians)
     /// </summary>
-    float Zoom { get; set; }
+    double Zoom { get; set; }
 
     /// <summary>
     /// If true, then we are in surface run mode, so special limits for zoom are applied
@@ -81,7 +81,7 @@ public interface ICamera
     /// <summary>
     /// Screen aspect ratio
     /// </summary>
-    float AspectRatio { get; set; }
+    double AspectRatio { get; set; }
 
     /// <summary>
     /// Camera up vector
@@ -101,12 +101,12 @@ public interface ICamera
     /// <summary>
     /// Zoom in
     /// </summary>
-    void ZoomIn(float steps);
+    void ZoomIn(double steps);
 
     /// <summary>
     /// Zoom out
     /// </summary>
-    void ZoomOut(float steps);
+    void ZoomOut(double steps);
 
     /// <summary>
     /// Project point from 3D space to viewport coordinates [-1; 1]
